Revise routing to avoid route to entity set controller if there is no related action
